Display only the first slide on page load
so that there’s no page jumping between page load and Unslider’s init.

Without this, the page would load all slides and display them, then
moments later when Unslider kicks in they’ll be floated and hidden,
causing the page flow to jump around. Doing this outside of Unslider
means it can’t properly detect the dimensions of the slides. This will
actually cause a few more reflows than would occur without this code
though.

This requires display: none; is set on .banner li:not(:first-of-type),
then running me.show() to set the CSS display property back to it’s
initial value in Unslider closer to when the slides are floated.