Fix missing last image / bleeding of next img
Changed the width on the slide (li) to percentage-based so that we no
longer get bleeding of the next/previous image, or the last image
wrapping to the next line (thus missing in the slider).

This happened when fluid is set to true. When the slider's parent
(.banner's parent) is a percentage-based width, that may calculate to a
decimal value (eg: 424.6343px) rather than a whole number. Unslider's
.banner will be 100% (so 424.2343px) but the slides themselves were
fixed width (424px), causing the wrap or bleed.